# optimize Float16.Mul