```
% benchstat .old.txt .new.txt
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/floats
cpu: Apple M1 Pro
               │  .old.txt   │              .new.txt               │
               │   sec/op    │   sec/op     vs base                │
Float16_Mul-10   4.377n ± 0%   3.444n ± 0%  -21.32% (p=0.000 n=10)

               │  .old.txt  │            .new.txt            │
               │    B/op    │    B/op     vs base            │
Float16_Mul-10   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

               │  .old.txt  │            .new.txt            │
               │ allocs/op  │ allocs/op   vs base            │
Float16_Mul-10   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
  - Improved multiplication for 16-bit floating-point numbers, now fully compliant with IEEE 754, including correct handling of special cases such as NaN, infinity, zero, underflow, and overflow.
- **Bug Fixes**
  - Enhanced accuracy and consistency for multiplication involving edge cases (e.g., infinity times zero, subnormal numbers).


<!-- end of auto-generated comment: release notes by coderabbit.ai -->